refactor: move DTOs
