+++ src/main/java/faang/school/projectservice/dto/moment/MomentRequestDto.java
@@ -1,7 +1,7 @@
-package faang.school.projectservice.dto.client.moment;
+package faang.school.projectservice.dto.moment;
 
 import faang.school.projectservice.validation.annotation.NullableNotBlank;
-import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -17,11 +17,8 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class MomentUpdateDto {
-    @NotNull
-    private Long id;
-
-    @NullableNotBlank
+public class MomentRequestDto {
+    @NotBlank
     @Size(max = 255)
     private String name;
 
@@ -29,7 +26,7 @@ public class MomentUpdateDto {
     private String description;
 
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
-    private LocalDateTime date;
+    private LocalDateTime date = LocalDateTime.now();
 
     @Size(min = 1)
     private List<@Positive Long> projectIds;
