+++ src/main/java/faang/school/projectservice/dto/moment/MomentUpdateDto.java
@@ -1,7 +1,7 @@
-package faang.school.projectservice.dto.client.moment;
+package faang.school.projectservice.dto.moment;
 
 import faang.school.projectservice.validation.annotation.NullableNotBlank;
-import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -17,8 +17,11 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class MomentRequestDto {
-    @NotBlank
+public class MomentUpdateDto {
+    @NotNull
+    private Long id;
+
+    @NullableNotBlank
     @Size(max = 255)
     private String name;
 
@@ -26,7 +29,7 @@ public class MomentRequestDto {
     private String description;
 
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
-    private LocalDateTime date = LocalDateTime.now();
+    private LocalDateTime date;
 
     @Size(min = 1)
     private List<@Positive Long> projectIds;
